requirements.txt
certifi==2018.10.15
Django==2.1.2
django-cors-headers==2.4.0
djangorestframework==3.8.2
elasticsearch==6.3.1
elasticsearch-dsl==6.2.1
ipaddress==1.0.22
pkg-resources==0.0.0
python-dateutil==2.7.3
pytz==2018.5
six==1.11.0
urllib3==1.23
